Fix vcr support file bug and filter sensitive data
The authorization in the headers is in an array instead of a string,
`sub` is a string method.

By using the `map` method this problem is solved.

In order to for the cassettes to work we need the correct URI but we
also need to filter out the `location_id` of the URI, by using the
`before_record` and `before_playback` hooks, this problem is solved.