Fix bug in Checkouts::Create service
As the `user_id` can be nil, calling `to_s` on it will send
an empty string.

When sending an empty string the API call will return an error,
as it needs this to create a customer on square.

Solved by adding an OR statement as there is always a guest token.