Update payment source
As we have to store the nonce to authorize the payment
when the order is finished,
I have added a nonce attribute on the payment source.

As we won't have the `square_order_id` that is saved as
the token on the payment_source,
I have removed the token validation on the payment source.