Add translation key for the payment method
As solidus will display the class name as payment method type in the
admin panel when creating a new payment method, it will just display
`Payment method` as the class name is PaymentMethod.

In order to make this more square specific, I have added a translation
key for that model.